requirements.txt
Flask
boto3
Flask-Cors
PyJWT
passlib
